+++ ceammc/ext/src/tests/test_ext_flow_change.cpp
@@ -11,6 +11,7 @@
  * contact the author of this file, or the owner of the project in which
  * this file belongs to.
  *****************************************************************************/
+#include "../base/function.h"
 #include "../flow/flow_change.h"
 #include "base_extension_test.h"
 #include "catch.hpp"
@@ -18,6 +19,7 @@
 
 #include <stdio.h>
 
+typedef TestExtension<Function> FunctionTest;
 typedef TestExtension<FlowChange> FlowChangeTest;
 
 static CanvasPtr cnv = PureData::instance().createTopCanvas("test_canvas");
@@ -29,6 +31,7 @@ TEST_CASE("flow.change", "[externals]")
         FlowChangeTest t("flow.change");
         REQUIRE(t.numInlets() == 1);
         REQUIRE(t.numOutlets() == 1);
+        REQUIRE_PROPERTY(t, @onrepeat, "");
     }
 
     SECTION("process")
@@ -111,5 +114,152 @@ TEST_CASE("flow.change", "[externals]")
 
         WHEN_SEND_LIST_TO(0, t, L4(1, 2, 3, 4));
         REQUIRE_LIST_AT_OUTLET(0, t, L4(1, 2, 3, 4));
+    }
+
+    SECTION("@onrepeat")
+    {
+        FunctionTest cb1("function", L1("test_callback1"));
+        FunctionTest cb2("function", L1("test_callback2"));
+
+#define STORE()                     \
+    {                               \
+        cb1.storeAllMessageCount(); \
+        cb2.storeAllMessageCount(); \
+    }
+
+        REQUIRE(Function::exists(gensym("test_callback1")));
+        REQUIRE(Function::exists(gensym("test_callback2")));
+
+        FlowChangeTest t("flow.change", L2("@onrepeat", "test_callback1"));
+        REQUIRE_PROPERTY(t, @onrepeat, gensym("test_callback1"));
+
+        SECTION("float")
+        {
+            // float
+            STORE();
+            WHEN_SEND_FLOAT_TO(0, t, 100);
+            REQUIRE_FLOAT_AT_OUTLET(0, t, 100);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb1);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb2);
+            // repeat
+            STORE();
+            WHEN_SEND_FLOAT_TO(0, t, 100);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(0, t);
+
+            REQUIRE_BANG_AT_OUTLET(1, cb1);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb2);
+
+            // repeat again
+            STORE();
+            WHEN_SEND_FLOAT_TO(0, t, 100);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(0, t);
+
+            REQUIRE_BANG_AT_OUTLET(1, cb1);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb2);
+
+            // new float
+            STORE();
+            WHEN_SEND_FLOAT_TO(0, t, 200);
+            REQUIRE_FLOAT_AT_OUTLET(0, t, 200);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb1);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb2);
+        }
+
+        SECTION("symbol")
+        {
+            STORE();
+            WHEN_SEND_SYMBOL_TO(0, t, "ABC");
+            REQUIRE_SYMBOL_AT_OUTLET(0, t, "ABC");
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb1);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb2);
+            // repeat
+            STORE();
+            WHEN_SEND_SYMBOL_TO(0, t, "ABC");
+            REQUIRE_NO_MESSAGES_AT_OUTLET(0, t);
+
+            REQUIRE_BANG_AT_OUTLET(1, cb1);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb2);
+
+            // repeat again
+            STORE();
+            WHEN_SEND_SYMBOL_TO(0, t, "ABC");
+            REQUIRE_NO_MESSAGES_AT_OUTLET(0, t);
+
+            REQUIRE_BANG_AT_OUTLET(1, cb1);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb2);
+
+            // new symbol
+            STORE();
+            WHEN_SEND_SYMBOL_TO(0, t, "ABCD");
+            REQUIRE_SYMBOL_AT_OUTLET(0, t, "ABCD");
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb1);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb2);
+        }
+
+        SECTION("list")
+        {
+            STORE();
+            WHEN_SEND_LIST_TO(0, t, L3(1, 2, 3));
+            REQUIRE_LIST_AT_OUTLET(0, t, L3(1, 2, 3));
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb1);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb2);
+            // change callback
+            t.setProperty("@onrepeat", L1("test_callback2"));
+            REQUIRE_PROPERTY(t, @onrepeat, gensym("test_callback2"));
+            // repeat
+            STORE();
+            WHEN_SEND_LIST_TO(0, t, L3(1, 2, 3));
+            REQUIRE_NO_MESSAGES_AT_OUTLET(0, t);
+
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb1);
+            REQUIRE_BANG_AT_OUTLET(1, cb2);
+
+            // repeat again
+            STORE();
+            WHEN_SEND_LIST_TO(0, t, L3(1, 2, 3));
+            REQUIRE_NO_MESSAGES_AT_OUTLET(0, t);
+
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb1);
+            REQUIRE_BANG_AT_OUTLET(1, cb2);
+
+            // new list
+            STORE();
+            WHEN_SEND_LIST_TO(0, t, L4(1, 2, 3, 4));
+            REQUIRE_LIST_AT_OUTLET(0, t, L4(1, 2, 3, 4));
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb1);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb2);
+        }
+
+        SECTION("any")
+        {
+            STORE();
+            WHEN_SEND_ANY_TO(t, "any1", L1(1));
+            REQUIRE_ANY_AT_OUTLET(0, t, L2("any1", 1));
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb1);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb2);
+
+            // repeat
+            STORE();
+            WHEN_SEND_ANY_TO(t, "any1", L1(1));
+            REQUIRE_NO_MESSAGES_AT_OUTLET(0, t);
+
+            REQUIRE_BANG_AT_OUTLET(1, cb1);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb2);
+
+            // repeat again
+            STORE();
+            WHEN_SEND_ANY_TO(t, "any1", L1(1));
+            REQUIRE_NO_MESSAGES_AT_OUTLET(0, t);
+
+            REQUIRE_BANG_AT_OUTLET(1, cb1);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb2);
+
+            // new list
+            STORE();
+            WHEN_SEND_ANY_TO(t, "any2", L1(1));
+            REQUIRE_ANY_AT_OUTLET(0, t, L2("any2", 1));
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb1);
+            REQUIRE_NO_MESSAGES_AT_OUTLET(1, cb2);
+        }
     }
 }