+++ ceammc/CHANGELOG.md
@@ -5,6 +5,7 @@
 - function - bang handling fixed
 - list.choice - fixed hang on empty list
 - list.choice @norepeat property added - for non repeated index choices
+- flow.change @onrepeat function callback added
 
 ### Added:
 - math.round~ added